Bug 1580087. Fix 32bit * 32bit integer multiplies so they don't overflow in imagelib. r=aosmond
Differential Revision: https://phabricator.services.mozilla.com/D45303

+++ image/ClippedImage.cpp
@@ -512,7 +512,8 @@ nsIntSize ClippedImage::OptimalImageSizeForDest(const gfxSize& aDest,
 
     // Determine the size we'd prefer to render the inner image at, and ask the
     // inner image what size we should actually use.
-    gfxSize desiredSize(imgWidth * scale.width, imgHeight * scale.height);
+    gfxSize desiredSize(double(imgWidth) * scale.width,
+                        double(imgHeight) * scale.height);
     nsIntSize innerDesiredSize = InnerImage()->OptimalImageSizeForDest(
         desiredSize, aWhichFrame, aSamplingFilter, aFlags);
